smb: client: Close all files marked for deferred close immediately
SMB client has a feature to delay the close of a file on the server,
expecting if an open call comes for the same file, the file doesn't
need to be re-opened. The time duration of the deferred close is set
by the mount option `closetimeo`.
This patch lets a user force close all files marked for deferred close,
across all session using a module parameter: close_all_deferred_closes.
By writing a non-zero value to:
/sys/module/cifs/parameters/close_all_deferred_closes, this feature
can be used.

Signed-off-by: Piyush Sachdeva <s.piyush1024@gmail.com>